refactor: textTranslation component move to separate folder
