+++ rss-puzzle/src/dirt/components/gameField/supportField/blocks/translation/textTranslation/textTranslation.ts
@@ -1,5 +1,5 @@
-import * as style from './translation-style.module.scss';
-import { TextTranslation, ITextTranslationOption, ITextTranslationStyleList } from '../../../../../modules/layout/gameField/supportField/blocks/translation/textTranslation';
+import * as style from './textTranslation-style.module.scss';
+import { TextTranslation, ITextTranslationOption, ITextTranslationStyleList } from '../../../../../../modules/layout/gameField/supportField/blocks/translation/textTranslation';
 
 const textTranslationStyleList: ITextTranslationStyleList =
 {
